requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-dotenv
google-genai
pydantic-settings